feat: modal component files separated
